Combine the functionality of the convert_to_tensors decorator and the dtypes constraint
