+++ tripy/tests/common/test_exception.py
@@ -23,7 +23,6 @@ from tests import helper
 
 import tripy as tp
 from tripy.common.exception import TripyException, _get_function_file_and_lines, str_from_stack_info, raise_error
-from tripy.frontend.utils import convert_to_tensors
 from tripy.utils import StackInfo, get_stack_info
 from tripy.utils.stack_info import SourceInfo
 
@@ -125,15 +124,14 @@ class TestRaiseError:
             in dedent(error_msg).strip()
         )
 
-    def test_convert_to_tensors_is_excluded(self):
-        filename, start_line, end_line = _get_function_file_and_lines(convert_to_tensors)
+    def test_constraints_is_excluded(self):
+        from tripy import constraints
+
         tensor = tp.ones((2, 3))
 
         stack_info = tensor.stack_info
 
-        assert any(
-            frame.file == filename and frame.line >= start_line and frame.line <= end_line for frame in stack_info
-        )
+        assert any(frame.module == constraints.__name__ for frame in stack_info)
 
         # Make sure that no extraneous wrapper code is included
         expected = dedent(
@@ -148,7 +146,7 @@ class TestRaiseError:
                 [0-9]+ |     return full\(shape, 1, dtype\)
                     |            ^^^^^^^^^^^^^^^^^^^^^ --- required from here
 
-                --> [a-z_/\.]+:[0-9]+ in test_convert_to_tensors_is_excluded\(\)
+                --> [a-z_/\.]+:[0-9]+ in test_constraints_is_excluded\(\)
                     |
                 [0-9]+ |         tensor = tp.ones\(\(2, 3\)\)
                     |                  ^^^^^^^^^^^^^^^ --- required from here
