+++ tripy/tripy/frontend/trace/ops/quantize.py
@@ -131,10 +131,11 @@ class Quantize(BaseTraceOp):
 
 
 @export.public_api(document_under="operations/quantization")
-@frontend_utils.convert_to_tensors(targets={"scale"})
 @constraints.dtypes(
     constraints={"input": "T1", "scale": "T1", "dtype": "T2", constraints.RETURN_VALUE: "T2"},
     variables={"T1": ["float32", "float16", "bfloat16"], "T2": ["int4", "int8", "float8"]},
+    convert_tensor_and_shape_likes=True,
+    conversion_targets={"scale"},
 )
 def quantize(
     input: "tripy.Tensor",
