+++ tripy/tripy/frontend/trace/ops/reshape.py
@@ -61,10 +61,11 @@ def infer_dimensions(input: "tripy.Tensor", shape: ShapeLike) -> ShapeLike:
 
 
 @export.public_api(document_under="operations/functions")
-@frontend_utils.convert_to_tensors(preprocess_args=infer_dimensions)
 @constraints.dtypes(
     constraints={"input": "T1", constraints.RETURN_VALUE: "T1"},
     variables={"T1": ["float32", "float16", "bfloat16", "float8", "int4", "int8", "int32", "int64", "bool"]},
+    convert_tensor_and_shape_likes=True,
+    conversion_preprocess_func=infer_dimensions,
 )
 def reshape(input: "tripy.Tensor", shape: ShapeLike) -> "tripy.Tensor":
     """
