+++ tripy/tripy/frontend/trace/ops/resize.py
@@ -105,10 +105,10 @@ def _check_mode(mode: str, align_corners: bool):
 
 
 @export.public_api(document_under="operations/functions")
-@frontend_utils.convert_to_tensors()
 @constraints.dtypes(
     constraints={"input": "T1", constraints.RETURN_VALUE: "T1"},
     variables={"T1": ["float32", "float16", "int8"]},
+    convert_tensor_and_shape_likes=True,
 )
 def resize(input: "tripy.Tensor", mode: str, output_shape: ShapeLike, align_corners: bool = False) -> "tripy.Tensor":
     r"""
