+++ tripy/tripy/frontend/utils.py
@@ -15,19 +15,14 @@
 # limitations under the License.
 #
 
-import functools
-import inspect
 from collections import deque
-from typing import Callable, Dict, List, Optional, Sequence, Set, Tuple, Union
+from typing import List, Union
 
-from tripy import utils
 from tripy.common.exception import raise_error
-from tripy.flat_ir.function import FlatIRFunction
 from tripy.flat_ir.ops import BaseFlatIROp
-from tripy.types import ShapeLike, TensorLike
 
 
-def tensor_from_shape_like(arg: ShapeLike) -> "tripy.Tensor":
+def tensor_from_shape_like(arg: "tripy.ShapeLike") -> "tripy.Tensor":
     from tripy.common.datatype import int32
     from tripy.frontend.dimension_size import DimensionSize
     from tripy.frontend.tensor import Tensor
@@ -67,201 +62,6 @@ def tensor_from_shape_like(arg: ShapeLike) -> "tripy.Tensor":
     # in a tensor with known shapes even though the new shape is actually known.
     out.trace_tensor.shape = [len(arg)]
     return out
-
-
-# Try to include correct column offsets for non-tensor arguments.
-def _add_column_info(arg, arg_index, is_kwarg, num_positional, func_name, arg_names):
-    from tripy import function_registry
-    from tripy.frontend.tensor import Tensor
-
-    assert isinstance(arg, Tensor), f"This function should only be called for objects that are already Tensor instances"
-
-    arg.stack_info.fetch_source_code()
-
-    # This is the stack depth in arg.stack_info where we find the decorated function.
-    for idx, source_info in enumerate(arg.stack_info):
-        if source_info.function == "wrapper" and source_info.module == __name__:
-            WRAPPER_STACK_DEPTH = idx + 1
-            break
-    else:
-        assert (
-            False
-        ), "`wrapper` function was not found in the call stack. Please update the check above if the name of the wrapper function has changed."
-
-    # Find the first caller of this function that is NOT the function registry.
-    # Also save the last dispatch target we see.
-    dispatch_target = None
-    for idx, source_info in enumerate(arg.stack_info[WRAPPER_STACK_DEPTH:]):
-        dispatch_target = source_info._dispatch_target or dispatch_target
-        if source_info.module not in utils.get_module_names_to_exclude_from_stack_info():
-            frame_index = idx + WRAPPER_STACK_DEPTH
-            break
-    else:
-        # Fallback path is just to look at the user code
-        frame_index = arg.stack_info.get_first_user_frame_index()
-        if frame_index is not None:
-            dispatch_target = arg.stack_info[frame_index - 1]._dispatch_target
-
-    # The function registry might prepend a class name to the dispatch target. We will strip it out here in order to match it.
-    if dispatch_target is not None and "." in dispatch_target:
-        dispatch_target = dispatch_target.split(".")[-1]
-
-    source_info = arg.stack_info[frame_index]
-
-    # The reverse binary ops need special handling since they will be swapped out for the non-reverse
-    # variants and the order of operands will be inverted.
-    REVERSE_BIN_OPS = {
-        "__radd__",
-        "__rmul__",
-        "__rsub__",
-        "__rpow__",
-        "__rtruediv__",
-    }
-
-    if dispatch_target in REVERSE_BIN_OPS:
-        assert arg_index in [0, 1]
-        arg_index = 0 if arg_index == 1 else 1
-        dispatch_target = dispatch_target.replace("__r", "__")
-
-    candidates = utils.get_arg_candidate_column_offsets(
-        source_info.code, arg_index, num_positional, dispatch_target or func_name, is_kwarg, arg_names
-    )
-
-    # Only set column range if there is exactly one candidate, otherwise we can't reliably determine
-    # what we should be pointing at.
-    if len(candidates) == 1:
-        source_info.column_range = candidates[0]
-
-
-# NOTE: Conversion to tensors needs to be done via a decorator so that we can add stack information
-# for non-tensors. Without having full context of the function signature, it is otherwise difficult to do so.
-def convert_to_tensors(targets: Set[str] = None, preprocess_args: Optional[Callable] = None):
-    """
-    Decorator that converts specified arguments to Tensors or DimensionSizes.
-    If the argument can be converted to a DimensionSize, it is. Otherwise, it is
-    converted to a Tensor.
-
-    If data type constraints are specified, any non-tensors converted will
-    also be casted to the expected data type, but will raise an exception for lossy
-    casts like float->int (but *not* for e.g. float32->float16).
-
-    Args:
-        targets: Names of arguments to convert to tensors. If not supplied, any arguments annotated
-            with `TensorLike` or `ShapeLike` are converted.
-
-        preprocess_args: A callback used to preprocess arguments before potential conversion. If provided,
-            this is always called, regardless of whether the decorator actually needed to perform conversion.
-            This will be called with all arguments that were passed to the decorated function and should
-            return a dictionary of all updated arguments. For a variadic arg, the dictionary entry for the name
-            should have a list of all the updated values.
-    """
-
-    def impl(func):
-        nonlocal targets
-
-        from tripy.constraints import TYPE_VERIFICATION
-
-        signature = inspect.signature(func)
-        targets = targets or {
-            name for name, param in signature.parameters.items() if param.annotation in {TensorLike, ShapeLike}
-        }
-        shape_likes = {name for name, param in signature.parameters.items() if param.annotation is ShapeLike}
-
-        constraints = {}
-        if func.__qualname__ in TYPE_VERIFICATION:
-            constraints = TYPE_VERIFICATION[func.__qualname__].constraints
-
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            from tripy.common.datatype import bool as tp_bool
-            from tripy.common.datatype import floating, integer
-            from tripy.constraints import get_arg_dtype
-            from tripy.frontend.dimension_size import DimensionSize
-            from tripy.frontend.tensor import Tensor
-            from tripy.frontend.trace.ops.cast import cast
-
-            all_args, var_arg_info = utils.merge_function_arguments(func, *args, **kwargs)
-
-            if preprocess_args is not None:
-
-                var_arg_name, var_arg_start_idx = utils.default(var_arg_info, (None, None))
-                new_args = preprocess_args(*args, **kwargs)
-                for index in range(len(all_args)):
-                    name, _ = all_args[index]
-                    if name in new_args:
-                        if name == var_arg_name:
-                            assert var_arg_start_idx is not None
-                            all_args[index] = (name, new_args[name][index - var_arg_start_idx])
-                        else:
-                            all_args[index] = (name, new_args[name])
-
-            # Materialize type variables from tensors.
-            type_vars = {}
-            for name, arg in all_args:
-                if name in constraints:
-                    dtype_result = get_arg_dtype(arg, func.__qualname__, name)
-                    if dtype_result:
-                        type_vars[constraints[name]] = dtype_result.value
-
-            new_args = []
-            new_kwargs = {}
-            for index, (name, arg) in enumerate(all_args):
-
-                def add_arg(arg):
-                    if name not in kwargs:
-                        new_args.append(arg)
-                    else:
-                        new_kwargs[name] = arg
-
-                if name in targets and not isinstance(arg, Tensor):
-                    if name in shape_likes:
-                        arg = tensor_from_shape_like(arg)
-                    else:
-                        # Python integers can always be casted to the most restrictive type, which is DimensionSize in Tripy.
-                        # DimensionSize can always be casted up to Tensor if needed, but the reverse is not true.
-                        # NOTE: We do not use isinstance here because bool is a subclass of int.
-                        arg = DimensionSize.create_directly(arg) if type(arg) is int else Tensor.create_directly(arg)
-
-                    _add_column_info(
-                        arg,
-                        index,
-                        name in kwargs,
-                        len(args),
-                        func.__name__,
-                        [name for name, _ in all_args],
-                    )
-
-                    dtype = None
-                    if name in constraints and constraints[name] in type_vars:
-                        dtype = type_vars[constraints[name]]
-
-                    if dtype is not None:
-                        # Refuse to do unsafe casts like float -> int.
-                        # NOTE: We do this check all the way down here so we have better stack information for `arg`.
-                        UNSAFE_CASTS = {floating: integer, integer: tp_bool}
-                        if any(
-                            issubclass(arg.dtype, frm) and issubclass(dtype, to) for frm, to in UNSAFE_CASTS.items()
-                        ):
-                            raise_error(
-                                f"Refusing to automatically cast '{arg.dtype}' argument to '{dtype}'.",
-                                [
-                                    f"Hint: You may want to manually cast other operands in this expression to compatible types.\n",
-                                    "Note: argument was: ",
-                                    arg,
-                                ],
-                            )
-
-                        original_stack_info = arg.stack_info
-                        arg = cast(arg, dtype=dtype)
-                        arg.stack_info = original_stack_info
-
-                add_arg(arg)
-
-            return func(*new_args, **new_kwargs)
-
-        return wrapper
-
-    return impl
 
 
 def topological_sort(ops: List[Union["BaseTraceOp", BaseFlatIROp]]) -> List[Union["BaseTraceOp", BaseFlatIROp]]:
